+++ openfisca_us/variables/hhs/tanf/cash/tanf_gross_unearned_income.py
@@ -9,6 +9,3 @@ class tanf_gross_unearned_income(Variable):
     documentation = "Gross unearned income for calculating Temporary Assistance for Needy Families benefit."
     unit = USD
     reference = "https://www.dhs.state.il.us/page.aspx?item=15814"
-
-    def formula(spm_unit, period, parameters):
-        return spm_unit("ssi", period)